Don't await_del if it's resolved
On Windows, there appears to be a bug in destructor ordering,
which causes the task to be destroyed prior to the awaiter. This
causes an SEH (ie, SEGV).

To work around this - and it may well be undefined territory - the
awaiter doesn't bother calling await_del if the awaiter has resolved,
or await_resume called already.